Reference

Contents

Reference

AutoGrad

KnetArray

Utilities

Convolution

Optimization

Initialization

AutoGrad (advanced)

Function Index

AutoGrad

AutoGrad.grad
AutoGrad.gradloss
AutoGrad.gradcheck

KnetArray

Knet.KnetArray — Type.

KnetArray(T, dims)
KnetArray(a::AbstractArray)
Array(k::KnetArray)

Container for GPU arrays that supports most of the AbstractArray interface. KnetArrays and Arrays can be converted to each other as shown above, which involves copying to and from the GPU memory. Important differences from the alternative CudaArray are: (1) a custom memory manager that minimizes the number of calls to the slow cudaMalloc by reusing already allocated but garbage collected GPU pointers. (2) a custom getindex that handles ranges such as a[5:10] as views with shared memory instead of copies.

Supported functions:

Array operations: cat, convert, copy, display, eachindex, elsize, eltype, endof, fill!, first, getindex, hcat, isempty, length, linearindexing, ndims, ones, pointer, rand!, reshape, setindex!, similar, size, stride, summary, vcat, vec, zeros

Math operators: (-), abs, abs2, acos, acosh, asin, asinh, atan, atanh, cbrt, ceil, cos, cosh, cospi, erf, erfc, erfcinv, erfcx, erfinv, exp, exp10, exp2, expm1, floor, log, log10, log1p, log2, round, sign, sin, sinh, sinpi, sqrt, tan, tanh, trunc

Broadcasting operators: (.*), (.+), (.-), (./), (.<), (.<=), (.!=), (.==), (.>), (.>=), (.^), max, min. (Only Array-Scalar and Array-Vector broadcasting are supported)

Reduction operators: countnz, maximum, minimum, prod, sum, sumabs, sumabs2, vecnorm. (Only Array->Scalar and Array->Vector reductions are supported)

Linear algebra: (*), axpy!, permutedims (only 2D and 3D), transpose

Knet extras: cpu2gpu, gpu2cpu, relu, sigm, invx, logp, logsumexp, conv4, pool, deconv4, unpool, mat, update!

source

Knet.cpu2gpu — Function.

Transfer from regular Array to KnetArray.

source

Knet.gpu2cpu — Function.

Transfer from KnetArray to regular Array.

source

Utilities

Knet.dir — Function.

Knet.dir(path...)

Construct a path relative to Knet root.

Example

julia> Knet.dir("examples","mnist.jl")
"/home/dyuret/.julia/v0.5/Knet/examples/mnist.jl"

source

Knet.gpu — Function.

gpu() returns the id of the active GPU device or -1 if none are active.

gpu(true) resets all GPU devices and activates the one with the most available memory.

gpu(false) resets and deactivates all GPU devices.

gpu(d::Int) activates the GPU device d if 0 <= d < gpuCount(), otherwise deactivates devices.

gpu(true/false) resets all devices. If there are any allocated KnetArrays their pointers will be left dangling. Thus gpu(true/false) should only be used during startup. If you want to suspend GPU use temporarily, use gpu(-1).

gpu(d::Int) does not reset the devices. You can select a previous device and find allocated memory preserved. However trying to operate on arrays of an inactive device will result in error.

source

Knet.logp — Function.

logp(x,[dims])

Treat entries in x as as unnormalized log probabilities and return normalized log probabilities.

dims is an optional argument, if not specified the normalization is over the whole x, otherwise the normalization is performed over the given dimensions. In particular, if x is a matrix, dims=1 normalizes columns of x and dims=2 normalizes rows of x.

source

Knet.logsumexp — Function.

logsumexp(x,[dims])

Compute log(sum(exp(x),dims)) in a numerically stable manner.

dims is an optional argument, if not specified the summation is over the whole x, otherwise the summation is performed over the given dimensions. In particular if x is a matrix, dims=1 sums columns of x and dims=2 sums rows of x.

source

Knet.invx — Function.

invx(x) = (1./x)

source

Knet.relu — Function.

relu(x) = max(0,x)

source

Knet.sigm — Function.

sigm(x) = (1./(1+exp(-x)))

source

Convolution

Knet.conv4 — Function.

conv4(w, x; kwargs...)

Execute convolutions or cross-correlations using filters specified with w over tensor x.

Currently 4 or 5 dimensional KnetArrays with Float32 or Float64 entries are supported. If w has dimensions (W1,W2,...,I,O) and x has dimensions (X1,X2,...,I,N), the result y will have dimensions (Y1,Y2,...,O,N) where

Yi=1+floor((Xi+2*padding[i]-Wi)/stride[i])

Here I is the number of input channels, O is the number of output channels, N is the number of instances, and Wi,Xi,Yi are spatial dimensions. padding and stride are keyword arguments that can be specified as a single number (in which case they apply to all dimensions), or an array/tuple with entries for each spatial dimension.

Keywords

padding=0: the number of extra zeros implicitly concatenated at the start and at the end of each dimension.

stride=1: the number of elements to slide to reach the next filtering window.

upscale=1: upscale factor for each dimension.

mode=0: 0 for convolution and 1 for cross-correlation.

alpha=1: can be used to scale the result.

algo=0: specifies which convolution algorithm shoud be used to compute the results. See the CUDNN User Guide for details.

workSpace=C_NULL: data pointer to GPU memory to a workspace needed to able to execute the specified algorithm.

workSpaceSizeInBytes=0: the size in bytes of the provided workSpace. Default=0.

handle: handle to a previously created cuDNN context. Defaults to a Knet allocated handle.

source

Knet.pool — Function.

pool(x; kwargs...)

Compute pooling of input values (i.e., the maximum or average of several adjacent values) to produce an output with smaller height and/or width.

Currently 4 or 5 dimensional KnetArrays with Float32 or Float64 entries are supported. If x has dimensions (X1,X2,...,I,N), the result y will have dimensions (Y1,Y2,...,I,N) where

Yi=1+floor((Xi+2*padding[i]-window[i])/stride[i])

Here I is the number of input channels, N is the number of instances, and Xi,Yi are spatial dimensions. window, padding and stride are keyword arguments that can be specified as a single number (in which case they apply to all dimensions), or an array/tuple with entries for each spatial dimension.

Keywords:

window=2: the pooling window size for each dimension.

padding=0: the number of extra zeros implicitly concatenated at the start and at the end of each dimension.

stride=window: the number of elements to slide to reach the next pooling window.

mode=0: 0 for max, 1 for average including padded values, 2 for average excluding padded values.

maxpoolingNanOpt=0: Nan numbers are not propagated if 0, they are propagated if 1.

alpha=1: can be used to scale the result.

handle: Handle to a previously created cuDNN context. Defaults to a Knet allocated handle.

source

Knet.mat — Function.

mat(x)

Reshape x into a two-dimensional matrix.

This is typically used when turning the output of a 4-D convolution result into a 2-D input for a fully connected layer. For 1-D inputs returns reshape(x, (length(x),1)). For inputs with more than two dimensions of size (X1,X2,...,XD), returns

reshape(x, (X1*X2*...*X[D-1],XD))

source

Knet.deconv4 — Function.

Deconvolution; reverse of convolution.

source

Knet.unpool — Function.

Unpooling; reverse of pooling.

source

Optimization

TODO: need blurb here about how optimization works, need to apply update! to individual weight arrays etc.

Knet.update! — Function.

update! takes a weight array w, the gradient (g) of the objective function w.r.t w and an instance of the parameters of an optimization method. It updates w depending on g and the specified optimization method with parameters in-place. If the any parameters of the optimization method need to be updated, it is also updated in-place. The function returns updated w and the instance of the parameters.

** Arguments **

w

g

prms::T

T might be one of the following:

Sgd
Momentum
Adagrad
Adadelta
Rmsprop
Adam

** Usage **

#prms = Sgd()
#prms = Momentum()
#prms = Adagrad()
#prms = Adadelta()
#prms = Rmsprop()
prms = Adam()

w, prms = update!(w, g, prms)
#update!(w, g, prm)

source

Knet.Sgd — Type.

Stochastic gradient descent is an optimization technique to minimize an objective function paremeterized by a model's parameters. It updates the parameters in the opposite direction of the gradient obtained by taking the gradient of the objective function w.r.t the paremeters. The learning rate (lr) determines the size of the step. It updates the weight with the following formula:

w = w - lr * g

where w is the weight, g is the gradient of the objective function w.r.t w and lr is the learning rate.

** Arguments **

lr::AbstractFloat=0.001

** Usage **

You can create the Sgd instance with default learning rate value or you can sepecify the learning rate. Then you can use the created instance in the update! function.

#prms = Sgd()# use the default value
prms = Sgd(;lr=0.1)
update!(w, g, prms)

#You can change the lr later
prms.lr = 0.01

source

Knet.Momentum — Type.

Momentum method tries to accelerate the Sgd by adding a velocity term to the update. It also decreases the oscilation between the opposite directions. It updates the weight with the following formulas:

velocity = gamma * velocity + lr * g
w = w - velocity

where w is the weight, g is the gradient of the objective function w.r.t w, lr is the learning rate, gamma is the momentum parameter, velocity is an array with the same size and type of w and holds the accelerated gradients.

** Arguments **

lr::AbstractFloat=0.001

gamma::AbstractFloat=0.9

velocity=zeros(w)

** Usage ** You can create the Momentum instance with default values or you can sepecify the parameters. Then you can use the created instance in the update! function.

#prms = Momentum() # use default values
prms = Momentum(;lr=0.1, gamma=0.95) # generally you do not need to specify the velocity parameter
update!(w, g, prms)

#You can change the parameters later
prms.lr=0.01
prms.gamma=0.9

Reference:

Qian, N. (1999). On the momentum term in gradient descent learning algorithms. Neural Networks : The Official Journal of the International Neural Network Society, 12(1), 145–151. http://doi.org/10.1016/S0893-6080(98)00116-6

source

Knet.Adagrad — Type.

Adagrad is one of the methods that adapts the learning rate to the parameters. It stores the sum of the squares of the gradients to scale the learning rate. The learning rate is adapted for each gradient by the value of current gradient divided by the accumulated gradients. Hence, the learning rate is greater for the parameters where the accumulated gradients are small and the learning rate is small if the accumulated gradients are big. It updates the weight with the following formulas:

G = G + g .^ 2
w = w - g .* lr ./ sqrt(G + eps)

where w is the weight, g is the gradient of the objective function w.r.t w, lr is the learning rate, G is an array with the same size and type of w and holds the sum of the squares of the gradients. eps is a small value to prevent the zero value on the denominator.

** Arguments **

lr::AbstractFloat=0.001

eps::AbstractFloat=1e-6

G=zeros(w)

** Usage ** You can create the Adagrad instance with default values or you can sepecify the parameters. Then you can use the created instance in the update! function.

#prms = Adagrad() # use default values
prms = Adagrad(;lr=0.1) # generally you do not need to specify the G and eps parameters
update!(w, g, prms)

#You can change the parameters later
prms.lr=0.01

Reference:

Duchi, J., Hazan, E., & Singer, Y. (2011). Adaptive Subgradient Methods for Online Learning and Stochastic Optimization. Journal of Machine Learning Research, 12, 2121–2159. Retrieved from http://jmlr.org/papers/v12/duchi11a.html

source

Knet.Adadelta — Type.

Adadelta is an extension to Adagrad to preventing the convergence of the learning rate to zero with increase of time. Adadelta uses two ideas from Momentum and Adagrad. It scales the learning rate based on the accumulated gradients and holds the acceleration term like Momentum. It updates the weight with the following formulas:

G = (1-rho) * g .^ 2 + rho * G
update = g .* sqrt(delta + eps) ./ sqrt(G + eps)
w = w - lr * update
delta = rho * delta + (1-rho) * update .^ 2

where w is the weight, g is the gradient of the objective function w.r.t w, lr is the learning rate, G is an array with the same size and type of w and holds the sum of the squares of the gradients. eps is a small value to prevent the zero value on the denominator. rho is the momentum parameter and delta is an array with the same size and type of w and holds the sum of the squared updates.

** Arguments **

lr::AbstractFloat=0.001

rho::AbstractFloat=0.9

eps::AbstractFloat=1e-6

G=zeros(w)

delta=zeros(w)

** Usage ** You can create the Adadelta instance with default values or you can sepecify the parameters. Then you can use the created instance in the update! function.

#prms = Adadelta() # use default values
prms = Adadelta(;lr=0.1, rho=0.8) # generally you do not need to specify the G, delta and eps parameters
update!(w, g, prms)

#You can change the parameters later
prms.lr=0.01
prms.rho=.9
prms.eps=1-e8

Reference:

Zeiler, M. D. (2012). ADADELTA: An Adaptive Learning Rate Method. Retrieved from http://arxiv.org/abs/1212.5701

source

Knet.Rmsprop — Type.

Rmsprop is a similar method to Adadelta that tries to improve Adagrad. It scales the learning rates by dividing the root mean squared of the gradients. It updates the weight with the following formula:

G = (1-rho) * g .^ 2 + rho * G
w = w - lr * g ./ sqrt(G + eps)

where w is the weight, g is the gradient of the objective function w.r.t w, lr is the learning rate, G is an array with the same size and type of w and holds the sum of the squares of the gradients. eps is a small value to prevent the zero value on the denominator. rho is the momentum parameter and delta is an array with the same size and type of w and holds the sum of the squared updates.

** Arguments **

lr::AbstractFloat=0.001

rho::AbstractFloat=0.9

eps::AbstractFloat=1e-6

G=zeros(w)

** Usage ** You can create the Rmsprop instance with default values or you can sepecify the parameters. Then you can use the created instance in the update! function.

#prms = Rmsprop() # use default values
prms = Rmsprop(;lr=0.1, rho=0.9) # generally you do not need to specify the G and eps parameters
update!(w, g, prms)

#You can change the parameters later
prms.lr=0.01
prms.rho=0.8
prms.eps=1-e8

Reference:

Tieleman, Tijmen, and Geoffrey Hinton. "Lecture 6.5-rmsprop: Divide the gradient by a running average of its recent magnitude." COURSERA: Neural Networks for Machine Learning 4.2 (2012).

source

Knet.Adam — Type.

Adam is one of the methods that compute the adaptive learning rate. It stores accumulated gradients (first moment) and the sum of the squared of gradients (second). It scales the first and second moment with increase of time. Here is the update formulas:

m = beta1 * m + (1 - beta1) * g
v = beta2 * v + (1 - beta2) * g .* g
mhat = m ./ (1 - beta1 ^ t)
vhat = v ./ (1 - beta2 ^ t)
w = w - (lr / (sqrt(vhat) + eps)) * mhat

where w is the weight, g is the gradient of the objective function w.r.t w, lr is the learning rate, m is an array with the same size and type of w and holds the accumulated gradients. v is an array with the same size and type of w and holds the sum of the squares of the gradients. eps is a small value to prevent the zero value on the denominator. beta1 and beta2 are the parameters to calculate bias corrected first and second moments. t is the update count.

** Arguments **

lr::AbstractFloat=0.001

beta1::AbstractFloat=0.9

beta2::AbstractFloat=0.999

t::AbstractFloat=1

eps::AbstractFloat=1e-8

fstm=zeros(w)

scndm=zeros(w)

** Usage ** You can create the Adam instance with default values or you can sepecify the parameters. Then you can use the created instance in the update! function.

#prms = Adam() # use default values
prms = Adam(;lr=0.1, beta1=0.95, beta2=0.995) # # generally you do not need to specify the fstm, scndm and eps parameters
update!(w, g, prms)

#You can change the parameters later
prms.lr=0.01
prms.beta1=0.8
prms.beta2=0.9
prms.eps=1e-9

Reference:

Kingma, D. P., & Ba, J. L. (2015). Adam: a Method for Stochastic Optimization. International Conference on Learning Representations, 1–13.

source

Initialization

Knet.gaussian — Function.

gaussian(a...; mean=0.0, std=0.01)

Return a Gaussian array with a given mean and standard deviation. The a arguments are passed to randn.

source

Knet.xavier — Function.

xavier(a...)

Xavier initialization. The a arguments are passed to rand. See (Glorot and Bengio 2010) for a description. Caffe implements this slightly differently. Lasagne calls it GlorotUniform.

source

Knet.bilinear — Function.

Bilinear interpolation filter weights; used for initializing deconvolution layers.

Adapted from https://github.com/shelhamer/fcn.berkeleyvision.org/blob/master/surgery.py#L33

Arguments:

T : Data Type

fw: Width upscale factor

fh: Height upscale factor

IN: Number of input filters

ON: Number of output filters

Example usage:

w = bilinear(Float32,2,2,128,128)

source

AutoGrad (advanced)

TODO: blurb here about how AutoGrad works, what the Rec type is etc.

AutoGrad.@primitive — Macro.

@primitive fx g1 g2... can be used to define a new primitive and (optionally) its gradients.

Julia supports multiple dispatch, i.e. a single function can have multiple methods with different arg types. AutoGrad supports multiple dispatch for primitives and gradients. Thus fx is a typed method declaration such as:

@primitive sin(x::Number)

@primitive hypot(x1::Array,x2::Array),dy,y

The second example specifies variable names for the output gradient dy and the output y after the method declaration which can be used in gradient expressions. Untyped, ellipsis and keyword arguments are ok as in f(a::Int,b,c...;d=1). Parametric methods such as f{T<:Number}(x::T) cannot be used.

The @primitive macro turns the first example into:

local sin_r = recorder(sin)
sin{T<:Number}(x::Rec{T}) = sin_r(x)

This will cause any call to sin with a Rec{T<:Number} argument to be recorded. With multiple arguments things are a bit more complicated. Here is what happens with the second example:

local hypot_r = recorder(hypot)
hypot{T<:Array,S<:Array}(x1::Rec{T},x2::Rec{S})=hypot_r(x1,x2)
hypot{T<:Array,S<:Array}(x1::Rec{T},x2::S)=hypot_r(x1,x2)
hypot{T<:Array,S<:Array}(x1::T,x2::Rec{S})=hypot_r(x1,x2)

We want the recorder version to be called if any one of the arguments is a boxed Rec. There is no easy way to specify this in Julia, so the macro generates all 2^N-1 boxed/unboxed argument combinations.

The method declaration can optionally be followed by gradient expressions. Here are the same examples with gradients:

@primitive sin(x::Number),dy (dy*cos(x))

@primitive hypot(x1::Array,x2::Array),dy,y (dy.*x1./y) (dy.*x2./y)

Note that the parameters, the return variable and the output gradient of the original function can be used in the gradient expressions.

In AutoGrad, gradients are defined using gradient methods that have the following signature:

f(Grad{i},dy,y,x...) => dx[i]

For the first example here is the generated gradient method:

sin{T<:Number}(::Type{Grad{1}}, dy, y, x::Rec{T})=(dy*cos(x))

For the second example a different gradient method is generated for each argument:

hypot{T<:Array,S<:Array}(::Type{Grad{1}},dy,y,x1::Rec{T},x2::Rec{S})=(dy.*x1./y) hypot{T<:Array,S<:Array}(::Type{Grad{2}},dy,y,x1::Rec{T},x2::Rec{S})=(dy.*x2./y)

In fact @primitive generates four more definitions for the other boxed/unboxed argument combinations.

Non-differentiable functions such as sign, and non-numeric functions such as size should be defined using the @zerograd macro instead.

source

AutoGrad.@zerograd — Macro.

@zerograd f(args...; kwargs...) allows f to handle its Rec inputs by unboxing them like @primitive, but unlike @primitive it does not record its actions or return a Rec result. Some functions, like sign(), have zero gradient. Others, like length() have discrete or constant outputs. These need to handle Rec inputs, but do not need to record anything and can return regular values. Their output can be treated like a constant in the program. Use the @zerograd macro for those. Note that kwargs are NOT unboxed.

source

AutoGrad.getval — Function.

getval(x) unboxes x if it is a Rec, otherwise returns x.

source

Function Index

Knet.Adadelta

Knet.Adagrad

Knet.Adam

Knet.KnetArray

Knet.Momentum

Knet.Rmsprop

Knet.Sgd

AutoGrad.getval

AutoGrad.grad

Knet.bilinear

Knet.conv4

Knet.cpu2gpu

Knet.deconv4

Knet.dir

Knet.gaussian

Knet.gpu

Knet.gpu2cpu

Knet.invx

Knet.logp

Knet.logsumexp

Knet.mat

Knet.pool

Knet.relu

Knet.sigm

Knet.unpool

Knet.update!

Knet.xavier

AutoGrad.@primitive

AutoGrad.@zerograd